Fix issue where order of CLI arguments matters
The TestCommand uses ignoreValidationErrors to allow it to accept and
pass through arbitrary options to phpunit/paratest. For example it does
not define `--stop-on-error` but can pass it through to phpunit.

A side effect of this is that the order of options matters, e.g.
`php artisan test -p --recreate-databases --stop-on-error` does recreate
databases, while
`php artisan test -p --stop-on-error --recreate-databases` does not.

This appears to be because when the validator hits an option it does not
know about, it stops processing and any subsequent options cannot be
resolved via $this->option.

It looks like this was previously handled when passing through options
to phpunit/paratest but was not handled when setting environment
variables, this change fixes that.